+++ notes-app/src/resetForm.js
@@ -1,0 +1,6 @@
+export const resetForm = () => {
+    textInput.value = "";
+    dateInput.value = "";
+    textarea.value = "";
+    category.value = "";
+};
