--- notes-app/src/someVar.js
@@ -1,1 +1,0 @@
-export const someVar = 5;
